Add files via upload
I have explore the case study on Danny's dinner as part of the Data with Danny case study challenge. I dive deep into the data and leverage basic and intermediate SQL functions to analyze the data and drive insights from the information. I showcase the power of data analysis in business decision-making and provide valuable insights.